feat: config webpack
