+++ index.html
@@ -12,7 +12,8 @@
 <body>
   <textarea id="text" cols="30" rows="10"></textarea>
   <div id="keyboard"></div>
-  <script type="module" src="./src/index.js"></script>
+  <!-- <script type="module" src="./src/index.js"></script> -->
+  <script src="dist/main.js"></script>
 </body>
