+++ package.json
@@ -2,6 +2,8 @@
   "devDependencies": {
     "eslint": "^8.39.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.27.5"
+    "eslint-plugin-import": "^2.27.5",
+    "webpack": "^5.81.0",
+    "webpack-cli": "^5.0.2"
   }
 }
